moved final testInfo calculations to context
