+++ src/components/TestResults/TestResults.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/jsx-key */
 import "../../styles/componentStyles/typing-test-results.scss";
 
-import React, { useEffect, useRef, useState} from "react";
+import React, { useEffect, useRef} from "react";
 import WpmGraph from "./WpmGraph";
 import { createScoreDocument } from "../../firebase/POST/scorePosts";
 import { useUserContext } from "../../contexts/UserContext";
@@ -9,8 +9,6 @@ import Statistics from "./Statistics";
 import HighScores from "./HighScores";
 import { updateUserSummary } from "../../firebase/POST/userPosts";
 import { useTestInformationContext } from "../../contexts/TestInformationContext";
-import { calculateAccuracy } from "../../functions/calculations/calculateAccuracy";
-import { calculateConsistency } from "../../functions/calculations/calculateConsistency";
 
 export interface TestResultsProps {
    
@@ -21,39 +19,11 @@ export interface TestResultsProps {
 
 const TestResults = ({resultsComponentOpacity, resultsComponentDisplay}: TestResultsProps ) => {
 	const {user, userDocument, isHeaderOpen, setIsHeaderOpen} = useUserContext();
-	const {testInformation, setTestInformation, showResultsComponent, isTestSubmitted, setIsTestSubmitted} = useTestInformationContext();
-
-	const [isCalculationsComplete, setIsCalculationsComplete] = useState<boolean>(false);
+	const {testInformation, showResultsComponent, isTestSubmitted, setIsTestSubmitted, isCalculationsComplete} = useTestInformationContext();
 
 	const isHeaderOpenRef = useRef<boolean>();
 	isHeaderOpenRef.current = isHeaderOpen;
-
-	useEffect(() => {
-		setIsCalculationsComplete(false);
-	}, []);
 	
-	// once results screen shown, calculate extra info to show 
-	useEffect(() => {
-		if (showResultsComponent) {	
-			setIsCalculationsComplete(false);
-
-			const accuracy = calculateAccuracy(testInformation);
-			const consistency = calculateConsistency(testInformation);	
-
-			setTestInformation({
-				...testInformation,
-				accuracy: accuracy,
-				consistency: consistency	
-			});	
-
-			setIsCalculationsComplete(true);
-		}
-		else {
-			setIsCalculationsComplete(false);
-			setIsTestSubmitted(localStorage.getItem("isSubmitted") === "true");
-		}
-	}, [showResultsComponent]);
-
 	// submit score when test hasnt already been submitted, user is logged in, and all result calculations are done
 	useEffect(() => {
 		if (!isTestSubmitted && showResultsComponent && isCalculationsComplete && user && userDocument)
