+++ src/contexts/TestInformationContext.tsx
@@ -4,6 +4,8 @@ import { TimedScoreDocument, WordCountScoreDocument } from "../firebase/firestor
 import { getHighScores } from "../firebase/GET/scoreGets";
 import { TestInformation } from "../interfaces/WordStructure";
 import { TestType } from "../enums";
+import { calculateAccuracy } from "../functions/calculations/calculateAccuracy";
+import { calculateConsistency } from "../functions/calculations/calculateConsistency";
 
 
 interface TestInformationContextProps {
@@ -13,8 +15,8 @@ interface TestInformationContextProps {
 	testInformation: TestInformation,
 	setTestInformation: React.Dispatch<React.SetStateAction<TestInformation>>,
 	showResultsComponent: boolean,
-	setShowResultsComponent: (bool: boolean) => void
-
+	setShowResultsComponent: (bool: boolean) => void,
+	isCalculationsComplete: boolean
 }
 
 const testInformationInitialState: TestInformation = {
@@ -39,7 +41,8 @@ export const TestInformationContext = createContext<TestInformationContextProps|
 	testInformation: testInformationInitialState,
 	setTestInformation: () => {},
 	showResultsComponent: false,
-	setShowResultsComponent: () => {}
+	setShowResultsComponent: () => {},
+	isCalculationsComplete: false
 });
 
 export const useTestInformationContext = () => {
@@ -55,12 +58,39 @@ export const TestInformationProvider = ({children}: any) => {
 	const [isTestSubmitted, setIsTestSubmitted] = useState<boolean>(localStorage.getItem("isSubmitted") === "true");
 	const [testInformation, setTestInformation] = useState<TestInformation>(testInformationInitialState);
 	const [showResultsComponent, setShowResultsComponent] = useState<boolean>(false);
+
+	const [isCalculationsComplete, setIsCalculationsComplete] = useState<boolean>(false);
+
 	
     
 	useEffect(() => {
 		console.log("TestInformationContext mounted");
 		setIsTestSubmitted(localStorage.getItem("isSubmitted") === "true");
 	}, []);
+
+	// once results screen shown, calculate final info for the TestInformation object
+	useEffect(() => {
+		if (showResultsComponent) {
+			console.log("calculating test results..");
+			setIsCalculationsComplete(false);
+	
+			const accuracy = calculateAccuracy(testInformation);
+			const consistency = calculateConsistency(testInformation);	
+
+			setTestInformation({
+				...testInformation,
+				accuracy: accuracy,
+				consistency: consistency	
+			});	
+
+			setIsCalculationsComplete(true);
+			console.log("test calculations done");
+		}
+		else {
+			setIsCalculationsComplete(false);
+			setIsTestSubmitted(localStorage.getItem("isSubmitted") === "true");
+		}
+	}, [showResultsComponent]);
 
 	const value: TestInformationContextProps = {
 		highScores,
@@ -69,7 +99,8 @@ export const TestInformationProvider = ({children}: any) => {
 		testInformation, 
 		setTestInformation,
 		showResultsComponent,
-		setShowResultsComponent
+		setShowResultsComponent,
+		isCalculationsComplete
 	};
 	
 	return (
